refactor cloud provider specific upload logic to separate classes; rename "data links" to "data-links" in plain text
