+++ src/main/java/io/seqera/tower/cli/commands/data/links/AddCmd.java
@@ -33,7 +33,7 @@ import java.util.Objects;
 
 @CommandLine.Command(
         name = "add",
-        description = "Add custom data link."
+        description = "Add custom data-link."
 )
 public class AddCmd extends AbstractApiCmd {
